+++ axum-macros/src/lib.rs
@@ -49,6 +49,7 @@ use syn::parse::Parse;
 
 mod debug_handler;
 mod from_request;
+mod typed_path;
 
 /// Derive an implementation of [`FromRequest`].
 ///
@@ -383,6 +384,16 @@ pub fn debug_handler(_attr: TokenStream, input: TokenStream) -> TokenStream {
 
     #[cfg(debug_assertions)]
     return expand_attr_with(_attr, input, debug_handler::expand);
+}
+
+/// Derive an implementation of [`axum_extra::routing::TypedPath`].
+///
+/// See that trait for more details.
+///
+/// [`axum_extra::routing::TypedPath`]: https://docs.rs/axum-extra/latest/axum_extra/routing/trait.TypedPath.html
+#[proc_macro_derive(TypedPath, attributes(typed_path))]
+pub fn derive_typed_path(input: TokenStream) -> TokenStream {
+    expand_with(input, typed_path::expand)
 }
 
 fn expand_with<F, I, K>(input: TokenStream, f: F) -> TokenStream