+++ axum-macros/tests/typed_path/fail/missing_field.rs
@@ -1,0 +1,9 @@
+use axum_macros::TypedPath;
+use serde::Deserialize;
+
+#[derive(TypedPath, Deserialize)]
+#[typed_path("/users/:id")]
+struct MyPath {}
+
+fn main() {
+}