+++ axum/src/extract/typed_header.rs
@@ -30,7 +30,6 @@ use std::ops::Deref;
 /// # axum::Server::bind(&"".parse().unwrap()).serve(app.into_make_service()).await.unwrap();
 /// # };
 /// ```
-#[cfg(feature = "headers")]
 #[cfg_attr(docsrs, doc(cfg(feature = "headers")))]
 #[derive(Debug, Clone, Copy)]
 pub struct TypedHeader<T>(pub T);
@@ -68,7 +67,6 @@ impl<T> Deref for TypedHeader<T> {
 
 /// Rejection used for [`TypedHeader`](super::TypedHeader).
 #[cfg(feature = "headers")]
-#[cfg_attr(docsrs, doc(cfg(feature = "headers")))]
 #[derive(Debug)]
 pub struct TypedHeaderRejection {
     name: &'static http::header::HeaderName,